The layermodel of the TOC has changed deeply.  Before this change, layers and layer collections were realizations from the interface ILayer. The user was able to construct trees of layers without deep limits.
Now, a LayerCollection can contain only ILayers, and LayerCollection is
not a realization of ILayer anymore. Consequently, the deep of the
resulting "tree" is limited to 2.

The DefaultMapContext has deeply change, too. There is no root node
anymore. This context now contain a list of IDisplayable, which are the
artifacts that can be shown in the TOC window.