Deep changes in the layer model. I am trying to fix all the problems occurring during tests.
The layermodel package is so modified. Some things are fixed, at least
partially : 

 * jaxb context can be partially retrieved
 * Layer can be added and their name can be changed

Remaining known problems : 

 * Removal of a layer doesn't work yet. The problem is linked to the
 listeners attached to the layer and to the mapcontext
 * Listeners are not, globally, well managed
 * There are still problems while working with jaxb contexts

Some of these problems are currently runtime problems. Exceptions are
thrown during tests execution. Most of them are linked to the fact that
it shall not be possible to acess layers of a mapcontext which is
closed.